arreglamos el llamado del DOM
