+++ src/lib/routes.js
@@ -11,7 +11,7 @@ export const routes = {
   '/wall': wall(),
 };
 const rootDiv = document.getElementById('root');
-rootDiv.innerHTML = routes[window.location.pathname];
+rootDiv.appendChild(routes[window.location.pathname]);
 
 const onNavigate = (pathname) => {
   window.history.pushState(
